[COMPILER] * BUG: The lowest long integer constant (-9223372036854775808) is correctly   interpreted now.
[GB.EVAL]
* BUG: The lowest long integer constant (-9223372036854775808) is correctly 
  interpreted now.


git-svn-id: svn://localhost/gambas/trunk@4230 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbc/gbc_trans.c
@@ -88,7 +88,7 @@ static bool read_integer(char *number, int base, bool minus, int64_t *result)
 			{
 				nbr2 = nbr * 10 + d;
 			
-				if ((nbr2 / 10) != nbr || nbr2 > (LLONG_MAX + minus))
+				if ((nbr2 / 10) != nbr || nbr2 > ((uint64_t)LLONG_MAX + minus))
 					return TRUE;
 			
 				nbr = nbr2;
